Morph Person class to Booking class, created BookingTest.java.
