+++ src/main/java/seedu/address/model/person/UniqueBookingList.java
@@ -1,0 +1,137 @@
+package seedu.address.model.person;
+
+import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
+
+import java.util.Iterator;
+import java.util.List;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import seedu.address.model.person.exceptions.DuplicateBookingException;
+import seedu.address.model.person.exceptions.BookingNotFoundException;
+import seedu.address.model.residence.Booking;
+
+/**
+ * A list of bookings that enforces uniqueness between its elements and does not allow nulls.
+ * A booking is considered unique by comparing using {@code Booking#isSameBooking(Booking)}. As such, adding and
+ * updating of bookings uses Booking#isSameBooking(Booking) for equality so as to ensure that the booking being added or
+ * updated is unique in terms of identity in the UniqueBookingList. However, the removal of a booking uses
+ * Booking#equals(Object) so as to ensure that the booking with exactly the same fields will be removed.
+ * Supports a minimal set of list operations.
+ *
+ * @see Booking#isSameBooking(Booking)
+ */
+public class UniqueBookingList implements Iterable<Booking> {
+
+    private final ObservableList<Booking> internalList = FXCollections.observableArrayList();
+    private final ObservableList<Booking> internalUnmodifiableList =
+            FXCollections.unmodifiableObservableList(internalList);
+
+    /**
+     * Returns true if the list contains an equivalent booking as the given argument.
+     */
+    public boolean contains(Booking toCheck) {
+        requireNonNull(toCheck);
+        return internalList.stream().anyMatch(toCheck::isSameBooking);
+    }
+
+    /**
+     * Adds a booking to the list.
+     * The booking must not already exist in the list.
+     */
+    public void add(Booking toAdd) {
+        requireNonNull(toAdd);
+        if (contains(toAdd)) {
+            throw new DuplicateBookingException();
+        }
+        internalList.add(toAdd);
+    }
+
+    /**
+     * Replaces the booking {@code target} in the list with {@code editedBooking}.
+     * {@code target} must exist in the list.
+     * The booking identity of {@code editedBooking} must not be the same as another existing booking in the list.
+     */
+    public void setBooking(Booking target, Booking editedBooking) {
+        requireAllNonNull(target, editedBooking);
+
+        int index = internalList.indexOf(target);
+        if (index == -1) {
+            throw new BookingNotFoundException();
+        }
+
+        if (!target.isSameBooking(editedBooking) && contains(editedBooking)) {
+            throw new DuplicateBookingException();
+        }
+
+        internalList.set(index, editedBooking);
+    }
+
+    /**
+     * Removes the equivalent booking from the list.
+     * The booking must exist in the list.
+     */
+    public void remove(Booking toRemove) {
+        requireNonNull(toRemove);
+        if (!internalList.remove(toRemove)) {
+            throw new BookingNotFoundException();
+        }
+    }
+
+    public void setBookings(UniqueBookingList replacement) {
+        requireNonNull(replacement);
+        internalList.setAll(replacement.internalList);
+    }
+
+    /**
+     * Replaces the contents of this list with {@code bookings}.
+     * {@code bookings} must not contain duplicate bookings.
+     */
+    public void setBookings(List<Booking> bookings) {
+        requireAllNonNull(bookings);
+        if (!bookingsAreUnique(bookings)) {
+            throw new DuplicateBookingException();
+        }
+
+        internalList.setAll(bookings);
+    }
+
+    /**
+     * Returns the backing list as an unmodifiable {@code ObservableList}.
+     */
+    public ObservableList<Booking> asUnmodifiableObservableList() {
+        return internalUnmodifiableList;
+    }
+
+    @Override
+    public Iterator<Booking> iterator() {
+        return internalList.iterator();
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof UniqueBookingList // instanceof handles nulls
+                        && internalList.equals(((UniqueBookingList) other).internalList));
+    }
+
+    @Override
+    public int hashCode() {
+        return internalList.hashCode();
+    }
+
+    /**
+     * Returns true if {@code bookings} contains only unique bookings.
+     */
+    private boolean bookingsAreUnique(List<Booking> bookings) {
+        for (int i = 0; i < bookings.size() - 1; i++) {
+            for (int j = i + 1; j < bookings.size(); j++) {
+                if (bookings.get(i).isSameBooking(bookings.get(j))) {
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
+}
