+++ src/main/java/seedu/address/model/person/exceptions/DuplicateBookingException.java
@@ -1,0 +1,11 @@
+package seedu.address.model.person.exceptions;
+
+/**
+ * Signals that the operation will result in duplicate Bookings (Bookings are considered duplicates if they have the
+ * same identity).
+ */
+public class DuplicateBookingException extends RuntimeException {
+    public DuplicateBookingException() {
+        super("Operation would result in duplicate bookings");
+    }
+}
