--- src/main/java/seedu/address/model/person/exceptions/DuplicatePersonException.java
@@ -1,11 +1,0 @@
-package seedu.address.model.person.exceptions;
-
-/**
- * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
- * identity).
- */
-public class DuplicatePersonException extends RuntimeException {
-    public DuplicatePersonException() {
-        super("Operation would result in duplicate persons");
-    }
-}
