+++ src/main/java/seedu/address/storage/JsonAdaptedBooking.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.residence.Booking;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
+import seedu.address.model.residence.Booking;
 
 /**
  * Json-friendly version of {@link Booking}.
