+++ src/test/java/seedu/address/model/person/UniquePersonListTest.java
@@ -60,23 +60,23 @@
 //
 //    @Test
 //    public void setPerson_nullTargetPerson_throwsNullPointerException() {
-//        assertThrows(NullPointerException.class, () -> uniquePersonList.setPerson(null, ALICE));
+//        assertThrows(NullPointerException.class, () -> uniquePersonList.setBooking(null, ALICE));
 //    }
 //
 //    @Test
 //    public void setPerson_nullEditedPerson_throwsNullPointerException() {
-//        assertThrows(NullPointerException.class, () -> uniquePersonList.setPerson(ALICE, null));
+//        assertThrows(NullPointerException.class, () -> uniquePersonList.setBooking(ALICE, null));
 //    }
 //
 //    @Test
 //    public void setPerson_targetPersonNotInList_throwsPersonNotFoundException() {
-//        assertThrows(PersonNotFoundException.class, () -> uniquePersonList.setPerson(ALICE, ALICE));
+//        assertThrows(PersonNotFoundException.class, () -> uniquePersonList.setBooking(ALICE, ALICE));
 //    }
 //
 //    @Test
 //    public void setPerson_editedPersonIsSamePerson_success() {
 //        uniquePersonList.add(ALICE);
-//        uniquePersonList.setPerson(ALICE, ALICE);
+//        uniquePersonList.setBooking(ALICE, ALICE);
 //        UniquePersonList expectedUniquePersonList = new UniquePersonList();
 //        expectedUniquePersonList.add(ALICE);
 //        assertEquals(expectedUniquePersonList, uniquePersonList);
@@ -87,7 +87,7 @@
 //        uniquePersonList.add(ALICE);
 //        Person editedAlice = new PersonBuilder(ALICE).withAddress(VALID_ADDRESS_BOB).withTags(VALID_TAG_HUSBAND)
 //                .build();
-//        uniquePersonList.setPerson(ALICE, editedAlice);
+//        uniquePersonList.setBooking(ALICE, editedAlice);
 //        UniquePersonList expectedUniquePersonList = new UniquePersonList();
 //        expectedUniquePersonList.add(editedAlice);
 //        assertEquals(expectedUniquePersonList, uniquePersonList);
@@ -96,7 +96,7 @@
 //    @Test
 //    public void setPerson_editedPersonHasDifferentIdentity_success() {
 //        uniquePersonList.add(ALICE);
-//        uniquePersonList.setPerson(ALICE, BOB);
+//        uniquePersonList.setBooking(ALICE, BOB);
 //        UniquePersonList expectedUniquePersonList = new UniquePersonList();
 //        expectedUniquePersonList.add(BOB);
 //        assertEquals(expectedUniquePersonList, uniquePersonList);
@@ -106,7 +106,7 @@
 //    public void setPerson_editedPersonHasNonUniqueIdentity_throwsDuplicatePersonException() {
 //        uniquePersonList.add(ALICE);
 //        uniquePersonList.add(BOB);
-//        assertThrows(DuplicatePersonException.class, () -> uniquePersonList.setPerson(ALICE, BOB));
+//        assertThrows(DuplicatePersonException.class, () -> uniquePersonList.setBooking(ALICE, BOB));
 //    }
 //
 //    @Test
@@ -129,7 +129,7 @@
 //
 //    @Test
 //    public void setPersons_nullUniquePersonList_throwsNullPointerException() {
-//        assertThrows(NullPointerException.class, () -> uniquePersonList.setPersons((UniquePersonList) null));
+//        assertThrows(NullPointerException.class, () -> uniquePersonList.setBookings((UniquePersonList) null));
 //    }
 //
 //    @Test
@@ -137,20 +137,20 @@
 //        uniquePersonList.add(ALICE);
 //        UniquePersonList expectedUniquePersonList = new UniquePersonList();
 //        expectedUniquePersonList.add(BOB);
-//        uniquePersonList.setPersons(expectedUniquePersonList);
+//        uniquePersonList.setBookings(expectedUniquePersonList);
 //        assertEquals(expectedUniquePersonList, uniquePersonList);
 //    }
 //
 //    @Test
 //    public void setPersons_nullList_throwsNullPointerException() {
-//        assertThrows(NullPointerException.class, () -> uniquePersonList.setPersons((List<Person>) null));
+//        assertThrows(NullPointerException.class, () -> uniquePersonList.setBookings((List<Person>) null));
 //    }
 //
 //    @Test
 //    public void setPersons_list_replacesOwnListWithProvidedList() {
 //        uniquePersonList.add(ALICE);
 //        List<Person> personList = Collections.singletonList(BOB);
-//        uniquePersonList.setPersons(personList);
+//        uniquePersonList.setBookings(personList);
 //        UniquePersonList expectedUniquePersonList = new UniquePersonList();
 //        expectedUniquePersonList.add(BOB);
 //        assertEquals(expectedUniquePersonList, uniquePersonList);
@@ -159,7 +159,7 @@
 //    @Test
 //    public void setPersons_listWithDuplicatePersons_throwsDuplicatePersonException() {
 //        List<Person> listWithDuplicatePersons = Arrays.asList(ALICE, ALICE);
-//        assertThrows(DuplicatePersonException.class, () -> uniquePersonList.setPersons(listWithDuplicatePersons));
+//        assertThrows(DuplicatePersonException.class, () -> uniquePersonList.setBookings(listWithDuplicatePersons));
 //    }
 //
 //    @Test
